MGSA-Student-Portal/requirements.txt

# REST API & Auth
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter==2.4.0
django==5.1.13
django-rest-knox>=5.0.0
# Static & Media
whitenoise>=6.0.0
Pillow>=10.0.0
cloudinary>=1.30.0
django-cloudinary-storage>=0.3.0

# Deployment
gunicorn>=20.0.0
dj-database-url>=1.0.0

# Excel & PDF Export
openpyxl>=3.1.2
reportlab>=4.0.7

# Testing & Faker
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
Faker>=18.0.0


